refactor: interface 관리 위한 Global Type d.ts파일 추가
