+++ client/src/types/type.d.ts
@@ -1,0 +1,8 @@
+declare module 'GlobalType' {
+  interface FriendsList {
+    idx: number;
+    userId: string;
+    username: string;
+    profileImg: string;
+  }
+}
